LayoutTests/imported/w3c/web-platform-tests/domxpath/predicates-expected.txt


FAIL An expression in a predicate should not change the context node assert_equals: expected 1 but got 0
